PySorterApp/requirements.txt
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiosignal==1.2.0
apparmor==3.1.1
appdirs==1.4.4
application-utility==1.3.2
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.1
bcrypt==4.0.0
beautifulsoup4==4.11.1
btrfsutil==5.19
cachetools==5.2.0
caffeine-ng==4.0.2
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==7.1.2
colorclass==2.2.2
colorlog==6.7.0
configobj==5.1.0.dev0
croniter==1.3.7
cryptography==37.0.4
Cython==0.29.32
distro==1.7.0
duplicity==0.8.23
ewmh==0.1.6
fasteners==0.17.3
faust==1.10.4
filelock==3.8.0
frozenlist==1.3.1
future==0.18.2
GNOME-AppFolders-Manager==0.4.6
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
gufw==22.4.0
httplib2==0.20.4
idna==3.3
inflect==6.0.0
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.9.1
kafka-python==1.4.7
layoutswitcherlib==0.8.27
LibAppArmor==3.1.1
libfdt==1.6.1
lit==14.0.6.dev0
manjaro-sdk==0.1
mode==4.3.2
more-itertools==8.13.0
multidict==6.0.2
mypy-extensions==0.4.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
oauth2client==4.1.3
opentracing==1.3.0
optimus-manager==1.4
ordered-set==4.1.0
packaging==21.3
pacman-mirrors==4.23.2
paramiko==2.11.0
Pillow==9.2.0
ply==3.11
protobuf==4.21.5
pulsectl==22.3.2
pwquality==1.4.4
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.21
pydantic==1.10.2
PyDrive2==1.14.0
PyGObject==3.42.2
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
pysmbc==1.0.23
python-dateutil==2.8.2
python-xlib==0.31
pytz==2022.1
pyxdg==0.28
PyYAML==6.0
ranger-fm==1.9.3
reportlab==3.6.11
requests==2.28.1
requests-file==1.5.1
robinhood-aiokafka==1.1.6
rsa==4.9
setproctitle==1.3.1
shiboken6==6.3.1
six==1.16.0
soupsieve==2.3.2.post1
systemd-python==235
team==1.0
terminaltables==3.1.10
tldextract==3.3.1
tomli==2.0.1
trove-classifiers==2022.8.31
typing_extensions==4.3.0
ufw==0.36.1
uritemplate==4.1.1
urllib3==1.26.12
validate==5.1.0.dev0
validate-pyproject==0.10.1
venusian==1.2.0
yarl==1.8.1
